MyRest_Api/MyRestApi/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
click==8.1.7
colorama==0.4.6
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
django-markdown==0.8.4
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
importlib-metadata==7.0.1
Markdown==3.5.2
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
zipp==3.17.0
